# Conflicts:
#	driver/web/adapter.go
#	driver/web/bbs_permission_policy.csv

+++ core/app_apis_bbs.go
@@ -25,46 +25,46 @@ import (
 	"github.com/rokwire/logging-library-go/v2/logs"
 )
 
-func (app *Application) bbsCreateMessage(inputMessage model.InputMessage) (*model.Message, error) {
+func (app *Application) bbsCreateMessages(inputMessages []model.InputMessage) ([]model.Message, error) {
 
-	return app.sharedCreateMessage(inputMessage)
+	return app.sharedCreateMessages(inputMessages)
 }
 
-func (app *Application) bbsDeleteMessage(l *logs.Log, serviceAccountID string, messageID string) error {
+func (app *Application) bbsDeleteMessages(l *logs.Log, serviceAccountID string, messagesIDs []string) error {
 	//in transaction
 	transaction := func(context storage.TransactionContext) error {
-		//find the message
-		message, err := app.storage.FindMessageWithContext(context, messageID)
+		//find the messages
+		messages, err := app.storage.FindMessagesWithContext(context, messagesIDs)
 		if err != nil {
 			return err
 		}
-		if message == nil {
-			return errors.New("no message for id - " + messageID)
+		if len(messagesIDs) != len(messages) {
+			return errors.New("not found message's")
 		}
+		/*
+			//validate if the service account is the sender of this message
+			valid := app.isSenderValid(serviceAccountID, *message)
+			if !valid {
+				return errors.New("not valid service account id for message - " + messageID)
+			}
 
-		//validate if the service account is the sender of this message
-		valid := app.isSenderValid(serviceAccountID, *message)
-		if !valid {
-			return errors.New("not valid service account id for message - " + messageID)
-		}
+			//delete the message
+			err = app.storage.DeleteMessageWithContext(context, message.OrgID, message.AppID, messageID)
+			if err != nil {
+				return err
+			}
 
-		//delete the message
-		err = app.storage.DeleteMessageWithContext(context, message.OrgID, message.AppID, messageID)
-		if err != nil {
-			return err
-		}
+			//delete the message recipients
+			err = app.storage.DeleteMessagesRecipientsForMessageWithContext(context, messageID)
+			if err != nil {
+				return err
+			}
 
-		//delete the message recipients
-		err = app.storage.DeleteMessagesRecipientsForMessageWithContext(context, messageID)
-		if err != nil {
-			return err
-		}
-
-		//delete the queue data items
-		err = app.storage.DeleteQueueDataForMessageWithContext(context, messageID)
-		if err != nil {
-			return err
-		}
+			//delete the queue data items
+			err = app.storage.DeleteQueueDataForMessageWithContext(context, messageID)
+			if err != nil {
+				return err
+			} */
 
 		return nil
 	}
